+++ library/oam_frame.c
@@ -27,7 +27,7 @@ void oam_build_eth_frame(uint8_t *dst_addr, uint8_t *src_addr, uint16_t type, ui
 }
 
 void oam_build_vlan_frame(uint8_t *dst_addr, uint8_t *src_addr, uint16_t tpi, uint8_t pcp, uint8_t dei,
-        uint16_t vlan_id, uint16_t ether_type, uint8_t* payload, uint32_t payload_s, uint8_t *frame)
+        uint16_t vlan_id, uint16_t ether_type, uint8_t* payload, size_t payload_s, uint8_t *frame)
 {
     struct oam_vlan_header vlan_hdr;
